engine mode second draft
