+++ packages/utils/src/utilities/index.ts
@@ -6,3 +6,4 @@ export * from './useDynamicDimension';
 export * from './useInterpolation';
 export * from './useItemsRange';
 export * from './useScreenDimensions';
+export * from './useEngine';
